requirements.txt
asgiref==3.10.0
Django==5.2.8
django-filter==25.2
djangorestframework==3.16.1
pillow==12.0.0
sqlparse==0.5.3
